backend/requirements.txt
alembic==1.11.0
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
factory-boy==3.2.1
Faker==18.9.0
fastapi==0.95.2
freezegun==1.2.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
